[UPDATE] Addition of parameter temporary_universal_model to the 'from_folder' method
This commit adds a new parameter temporary_universal_model to the from_folder method of the GapFiller class. If this parameter is set as True, a temporary universal model will be created that contains only the pathways related to the targets. This is useful for reducing the size of the universal model, improving the time taken when applying gap-filling algorithms.