[ASS] test for 'from_folder' method
This commit adds a test case for the GapFiller.from_folder() method with the temporary_universal_model parameter set to True. The test checks the correct creation of a temporary universal model and validates the presence of the expected model file in the specified folder. The test also ensures that the universal model is not None after the GapFiller object is created.