[ADD] Several new methods were added and from_folder method was updated
Updated method:
from_folder, cloned the model and set it as the universal_model_compartmentalized attribute if the compartmentalized file was found and the clone flag is True.

Added method: met_is_being_consumed_or_produced_at_reaction to check if a metabolite is being consumed or produced at a reaction.

Added method: check_metabolite_in_reaction to check if a metabolite is being consumed or produced at a reaction.
Iterated over the reactions and stored the involvement of the metabolite in each reaction.
Added print statements to display reaction information.

Added method:
search_dead_end to search for reactions in the universal model where a given metabolite is consumed or produced.
Implemented the functionality to search for consumed and/or produced reactions based on the provided parameters.
Added an optional verbose flag to print the reactions where the metabolite is consumed or produced.
Returned a tuple containing lists of consumed and produced reactions.

Added method:
 deal_dead_ends to handle dead-end metabolites.
Implemented the functionality to search for reactions in the universal model where the dead-end metabolite is consumed or produced.
Added logic to add the identified reactions to the draft model.
Handled the case where the dead-end metabolite is in different compartments by creating a transport reaction.
Updated the draft model with the added reactions.
Implemented the functionality to automatically identify dead-end metabolites if not provided.
Returned the updated draft model.